requirements.txt
flask
flask-socketio 
gevent
gevent-websocket
gunicorn
humanfriendly
hvac
psutil
pyjwt
pyyaml
requests_toolbelt
tqdm
websocket-client
xmltodict
zeroconf
# flask-server-timing
redis 
psycopg2-binary